+++ src/graph/flaps.js
@@ -1,116 +1,82 @@
-// /**
-//  * Rabbit Ear (c) Kraft
-//  */
-// import {
-// 	EPSILON,
-// } from "../math/constant.js";
-// import {
-// 	excludeS,
-// } from "../math/compare.js";
-// import {
-// 	cross2,
-// 	subtract2,
-// } from "../math/vector.js";
-// import {
-// 	getEdgesLineIntersection,
-// 	getEdgesCollinearToLine,
-// } from "./intersect/edges.js";
-// import {
-// 	makeFacesEdgesFromVertices,
-// } from "./make/facesEdges.js";
-// import {
-// 	makeFacesNormal,
-// } from "./normals.js";
-// import {
-// 	faceOrdersSubset,
-// 	linearizeFaceOrders,
-// } from "./orders.js";
+/**
+ * Rabbit Ear (c) Kraft
+ */
+import {
+	EPSILON,
+} from "../math/constant.js";
+import {
+	cross2,
+	subtract2,
+} from "../math/vector.js";
+import {
+	clone,
+} from "../general/clone.js";
+import {
+	invertArrayToFlatMap,
+	invertFlatToArrayMap,
+} from "./maps.js";
+import {
+	subgraphWithFaces,
+} from "./subgraph.js";
+import {
+	connectedComponents,
+} from "./connectedComponents.js";
+import {
+	makeFacesFaces,
+} from "./make/facesFaces.js";
+import {
+	makeFacesCenter2DQuick,
+} from "./make/faces.js";
+import {
+	foldLine,
+} from "./fold/foldGraph.js";
 
-// /**
-//  * @param {FOLD} graph a FOLD object
-//  * @param {VecLine2} line
-//  * @param {number} [epsilon=1e-6] an optional epsilon
-//  */
-// export const getEdgesSide = ({ vertices_coords, edges_vertices }, line, epsilon = EPSILON) => {
-// 	/** @param {[number, number]} edge_vertices @returns {number} */
-// 	const edgeSide = (edge_vertices) => edge_vertices
-// 		.map(v => vertices_coords[v])
-// 		.map(coord => subtract2(coord, line.origin))
-// 		.map(vec => cross2(line.vector, vec))
-// 		.sort((a, b) => Math.abs(b) - Math.abs(a))
-// 		.map(Math.sign)
-// 		.shift();
-// 	const edgesIntersection = getEdgesLineIntersection({
-// 		vertices_coords, edges_vertices,
-// 	}, line, epsilon, excludeS);
-// 	const edgesCollinear = {};
-// 	getEdgesCollinearToLine({ vertices_coords, edges_vertices }, line, epsilon)
-// 		.forEach(e => { edgesCollinear[e] = true; });
-// 	// -1: left, +1: right (todo check these), 0: both, 2: collinear (neither side)
-// 	return edges_vertices.map((edge_vertices, e) => {
-// 		if (edgesCollinear[e] === true) { return 2; }
-// 		if (edgesIntersection[e].point !== undefined) { return 0; }
-// 		return edgeSide(edge_vertices);
-// 	});
-// };
+/**
+ * @description Divide a folded graph with a line and return lists of connected
+ * faces (flaps). The result is returned as two sets representing either side
+ * of the line, inside each set is a list of connected faces arrays.
+ * Currently, flat folded crease patterns only (2D)
+ * @param {FOLD} graph a FOLD object
+ * @param {VecLine2} line
+ * @param {[number, number][]} [vertices_coordsFolded]
+ * @param {number} [epsilon=1e-6] an optional epsilon
+ * @returns {[ number[][], number[][] ]} left and right sides, each side
+ * contains a list of lists of faces, groups of connected sets of faces (flaps).
+ */
+export const getFlaps = ({
+	vertices_coords, edges_vertices, faces_vertices,
+}, line, vertices_coordsFolded = undefined, epsilon = EPSILON) => {
+	const graph = clone({ vertices_coords, edges_vertices, faces_vertices });
+	const {
+		vertices,
+		faces: { map }
+	} = foldLine(graph, line, { assignment: "F", vertices_coordsFolded }, epsilon);
+	const backmap = invertArrayToFlatMap(map);
 
-// /**
-//  * @param {FOLD} graph a FOLD object
-//  * @param {VecLine2} line
-//  * @param {number} [epsilon=1e-6] an optional epsilon
-//  */
-// export const getFacesSide = ({
-// 	vertices_coords, edges_vertices, faces_vertices, faces_edges,
-// }, line, epsilon = EPSILON) => {
-// 	if (!faces_edges) {
-// 		faces_edges = makeFacesEdgesFromVertices({ edges_vertices, faces_vertices });
-// 	}
-// 	const edgesSide = getEdgesSide({ vertices_coords, edges_vertices }, line, epsilon);
-// 	// filter out "collinear" edges, they don't matter here
-// 	const facesEdgesSide = faces_edges
-// 		.map(edges => edges
-// 			.map(e => edgesSide[e])
-// 			.filter(side => side !== 2));
-// 	const facesOverlapLine = facesEdgesSide
-// 		.map(sides => sides.includes(0));
-// 	const facesEdgesSameSide = facesEdgesSide
-// 		.map((sides, i) => (facesOverlapLine[i]
-// 			? false
-// 			: sides.reduce((a, b) => a && (b === sides[0]), true)));
-// 	return facesEdgesSameSide
-// 		.map((sameSide, f) => (sameSide ? facesEdgesSide[f][0] : 0));
-// };
+	const folded = { ...graph, vertices_coords: vertices.folded };
+	const faces_faces = makeFacesFaces(graph);
+	const faces_center = makeFacesCenter2DQuick(folded);
+	const faces_side = faces_center
+		.map(point => subtract2(point, line.origin))
+		.map(vector => cross2(vector, line.vector))
+		.map(Math.sign)
+		.map(s => (s + 1) / 2); // convert to 0, 1
+	const sidesFaces = invertFlatToArrayMap(faces_side);
+	// two subgraphs, each containing only the faces from either side of the line
+	const sidesGraphs = sidesFaces
+		.map(faces => subgraphWithFaces({ faces_faces }, faces));
 
-// /**
-//  * @description flat folded crease patterns only (2D)
-//  * @algorithm two approaches.
-//  * split the graph with a line (create a copy in memory), rebuild a new
-//  * faces_faces array, and save the faces-map, separate all faces in the graph
-//  * based on which side of the line
-//  * alternatively (without modifying the graph):
-//  *
-//  */
-// export const getFlapsThroughLine = ({
-// 	vertices_coords, edges_vertices, faces_vertices, faces_edges, faceOrders,
-// }, line, epsilon = EPSILON) => {
-// 	if (!faceOrders) { throw new Error("faceOrders required"); }
-// 	// for every face, get the intersection
-// 	const facesSide = getFacesSide({
-// 		vertices_coords, edges_vertices, faces_vertices, faces_edges,
-// 	}, line, epsilon);
-// 	const sidesFaces = [-1, 1]
-// 		.map(side => facesSide
-// 			.map((s, f) => ({ s, f }))
-// 			.filter(el => el.s === side || el.s === 0)
-// 			.map(el => el.f));
-// 	const sidesFaceOrders = sidesFaces
-// 		.map(faces => faceOrdersSubset(faceOrders, faces));
-// 	console.log("facesSide", facesSide);
-// 	console.log("sidesFaces", sidesFaces);
-// 	console.log("sidesFaceOrders", sidesFaceOrders);
-// 	const faces_normal = makeFacesNormal({ vertices_coords, faces_vertices });
-// 	const sidesLayersFace = sidesFaceOrders.map(orders => linearizeFaceOrders({
-// 		faceOrders: orders, faces_normal,
-// 	}));
-// 	console.log("sidesLayersFace", sidesLayersFace);
-// };
+	// convert each side into connected components (this first comes in as an
+	// inverted result, faces_set, convert it to sets_faces)
+	const sidesConnectedFaces = sidesGraphs
+		.map(({ faces_faces }) => connectedComponents(faces_faces))
+		.map(invertFlatToArrayMap);
+
+	// console.log("map", map);
+	// console.log("backmap", backmap);
+	const [sideA, sideB] = sidesConnectedFaces
+		.map(connectedFaces => connectedFaces
+			.map(faces => faces
+				.map(face => backmap[face])));
+	return [sideA, sideB];
+};