+++ src/graph/index.js
@@ -38,7 +38,7 @@ import * as cycles from "./cycles.js";
 import * as directedGraph from "./directedGraph.js";
 import * as disjoint from "./disjoint.js";
 import * as explodeMethods from "./explode.js";
-// import * as flaps from "./flaps.js";
+import * as flaps from "./flaps.js";
 import * as intersect from "./intersect.js";
 import * as join from "./join.js";
 import * as maps from "./maps.js";
@@ -107,7 +107,7 @@ const graphMethods = {
 	...directedGraph,
 	...disjoint,
 	...explodeMethods,
-	// ...flaps,
+	...flaps,
 	...intersect,
 	...join,
 	...make,